+++ packages/babel-plugin-nornj-in-jsx/src/exTag.js
@@ -1,3 +1,4 @@
+const nj = require('nornj/dist/nornj.common').default;
 const astUtil = require('./util/ast');
 const generate = require('./util/generate');
 const utils = require('./util/utils');
@@ -12,7 +13,7 @@ module.exports = function (babel) {
     if (astUtil.hasExPrefix(elName)) {
       elName = elName.substr(2);
     }
-    elName = utils.lowerFirst(elName);
+    elName = nj.lowerFirst(elName);
     const key = astUtil.getKey(node);
     const attrs = astUtil.getAttributeMap(node);
     const children = astUtil.getChildren(types, node);
@@ -53,7 +54,7 @@ module.exports = function (babel) {
           if (astUtil.hasExPrefix(subElName)) {
             subElName = subElName.substr(2);
           }
-          subElName = utils.lowerFirst(subElName);
+          subElName = nj.lowerFirst(subElName);
           const subAttrs = astUtil.getAttributeMap(subExTagNode);
           const subNewContextData = {};
           lastAttrStr = generate.buildAttrs(types, subElName, subAttrs, quasis, expressions, lastAttrStr, subNewContextData);